requirements.txt
tensorflow
torch
torchvision
transformers
librosa
numpy
pandas
matplotlib
Pillow